+++ project_test/tests/003_fields.py
@@ -9,6 +9,7 @@ from utils import read_output, write_output
 
 def test_inlinefield(output_test_path, rendered_template, helper, client):
     form = BasicInputForm()
+    pack = helper.template_pack
 
     helper.layout = Layout(
         InlineField('simple', label_column='large-7', input_column='large-5',
@@ -17,14 +18,15 @@ def test_inlinefield(output_test_path, rendered_template, helper, client):
 
     rendered = rendered_template(form, helper=helper)
 
-    attempted = read_output(output_test_path, "test_inlinefield.html")
-    #write_output(output_test_path, "test_inlinefield.html", rendered)
+    attempted = read_output(output_test_path, pack, "test_inlinefield.html")
+    #write_output(output_test_path, pack, "test_inlinefield.html", rendered)
 
     assert rendered == attempted
 
 
 def test_inlineswitchfield(output_test_path, rendered_template, helper, client):
     form = BoolInputForm()
+    pack = helper.template_pack
 
     helper.layout = Layout(
         InlineSwitchField('opt_in', label_column='large-8',
@@ -34,7 +36,7 @@ def test_inlineswitchfield(output_test_path, rendered_template, helper, client):
 
     rendered = rendered_template(form, helper=helper)
 
-    attempted = read_output(output_test_path, "test_inlineswitchfield.html")
-    #write_output(output_test_path, "test_inlineswitchfield.html", rendered)
+    attempted = read_output(output_test_path, pack, "test_inlineswitchfield.html")
+    #write_output(output_test_path, pack, "test_inlineswitchfield.html", rendered)
 
     assert rendered == attempted