+++ project_test/tests/utils.py
@@ -2,11 +2,11 @@ import os
 import io
 
 
-def read_output(filepath, filename):
+def read_output(filepath, pack, filename):
     """
     Read file and return its content
     """
-    destination = os.path.join(filepath, filename)
+    destination = os.path.join(filepath, pack, filename)
 
     with io.open(destination, 'r', encoding='utf-8') as f:
         content = f.read()
@@ -14,15 +14,15 @@ def read_output(filepath, filename):
     return content
 
 
-def write_output(filepath, filename, content):
+def write_output(filepath, pack, filename, content):
     """
     Write content to filepath+filename, create filepath if it does not
     allready exists
     """
-    if filepath and filepath != '.' and not os.path.exists(filepath):
-        os.makedirs(filepath)
+    if filepath and filepath != '.' and not os.path.exists(os.path.join(filepath, pack)):
+        os.makedirs(os.path.join(filepath, pack))
 
-    destination = os.path.join(filepath, filename)
+    destination = os.path.join(filepath, pack, filename)
 
     with io.open(destination, 'w', encoding='utf-8') as f:
         f.write(content)